+++ src/index.js
@@ -10,6 +10,9 @@ import configureStore from './store/configureStore';
 require('./favicon.ico'); // Tell webpack to load favicon.ico
 import './styles/styles.scss'; // Yep, that's right. You can import SASS/CSS files too! Webpack will run the associated loader and plug this into the page.
 import {syncHistoryWithStore} from 'react-router-redux';
+import injectTapEventPlugin from 'react-tap-event-plugin';
+
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 
 const store = configureStore();
 
@@ -17,9 +20,17 @@ const store = configureStore();
 const history = syncHistoryWithStore(browserHistory, store);
 
 
+try {
+	injectTapEventPlugin();
+} catch(error) {
+	console.log(error);
+}
+
 render(
 	<AppContainer>
-		<Root store={store} history={history}/>
+		<MuiThemeProvider>
+			<Root store={store} history={history}/>
+		</MuiThemeProvider>
 	</AppContainer>,
 	document.getElementById('app')
 );